+++ packages/dt_for_itables/src/index.d.ts
@@ -1,0 +1,1 @@
+declare module 'dt_for_itables';